requirements.txt
fastapi
uvicorn
pydantic
httpx
redis
asyncpg
psycopg2-binary
python-dotenv

openai
anthropic
google-generativeai
google-genai

numpy
python-dotenv

pgvector
sqlalchemy